@lochjin 2023-10-13
> Try moving the `b.ResetTimer()` after `New` otherwise you are benchmarking the initialization as well (which is not indicative of the write performance). In the read tests you already do this

Even after such modifications, there is still a significant difference