@MariusVanDerWijden 2023-10-13
It looks okay to me, the individual pebble write might be slower, but pebble is still significantly faster due to faster compaction when used in the actual node
```
BenchmarkIOPebbleDB/Write/WriteSorted-8         	1000000000	         0.006687 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/Write/WriteRandom-8         	1000000000	         0.006880 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/Read/ReadSorted-8           	1000000000	         0.0007375 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/Read/ReadRandom-8           	1000000000	         0.001268 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/Iteration/IterationSorted-8 	1000000000	         0.0008966 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/Iteration/IterationRandom-8 	1000000000	         0.0008488 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/BatchWrite/BenchWriteSorted-8         	1000000000	         0.001228 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/BatchWrite/BenchWriteRandom-8         	1000000000	         0.001827 ns/op	       0 B/op	       0 allocs/op
PASS
ok  	github.com/ethereum/go-ethereum/ethdb/pebble	0.440s
```

```

BenchmarkIOLevelDB/Write/WriteSorted-8         	1000000000	         0.002377 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/Write/WriteRandom-8         	1000000000	         0.003358 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/Read/ReadSorted-8           	1000000000	         0.0006967 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/Read/ReadRandom-8           	1000000000	         0.001898 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/Iteration/IterationSorted-8 	1000000000	         0.001150 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/Iteration/IterationRandom-8 	1000000000	         0.002078 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/BatchWrite/BenchWriteSorted-8         	1000000000	         0.008206 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/BatchWrite/BenchWriteRandom-8         	1000000000	         0.009791 ns/op	       0 B/op	       0 allocs/op
```