@holiman 2023-10-13
Also, `geth` almost never does a standalone write, but always batch-writes. 
```
BenchmarkIOPebbleDB/BatchWrite/BenchWriteSorted-8         	1000000000	         0.001228 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOPebbleDB/BatchWrite/BenchWriteRandom-8         	1000000000	         0.001827 ns/op	       0 B/op	       0 allocs/op

BenchmarkIOLevelDB/BatchWrite/BenchWriteSorted-8         	1000000000	         0.008206 ns/op	       0 B/op	       0 allocs/op
BenchmarkIOLevelDB/BatchWrite/BenchWriteRandom-8         	1000000000	         0.009791 ns/op	       0 B/op	       0 allocs/op
```

So even if we were to consider the synthetic benchmark important (which we don't, really, as @karalabe explained, since benchmarks do not capture the real-world performance on real-world payloads), it does point to a win for Pebble. 

Closing this, think it's nothing more to answer